feat(DeployedForm, TemplateBase): move utlity method to new js file from template base
