+++ frontend/src/components/forms/TemplateBase.vue
@@ -66,58 +66,6 @@ export default {
     }
   },
   methods: {
-    /**
-     * Convert form variables from service response format to form-js format
-     * @param {Object} formData - Raw form data from service response
-     * @returns {Object} Converted form data for form-js
-     * @example
-     * // Input (service response format):
-     * {
-     *   "availableOptions": {
-     *     "name": null,
-     *     "type": "Json", 
-     *     "value": "[{\"label\":\"Option A\",\"value\":\"a\"},{\"label\":\"Option B\",\"value\":\"b\"}]",
-     *     "valueInfo": {}
-     *   },
-     *   "userName": {
-     *     "name": "user_name",
-     *     "type": "String",
-     *     "value": "John Doe",
-     *     "valueInfo": {}
-     *   }
-     * }
-     * 
-     * // Output (form-js format):
-     * {
-     *   "availableOptions": [
-     *     {"label": "Option A", "value": "a"},
-     *     {"label": "Option B", "value": "b"}
-     *   ],
-     *   "user_name": "John Doe"
-     * }
-     */
-    convertFormDataForFormJs: function(formData) {
-      const convertedFormData = {}
-      Object.keys(formData).forEach(key => {
-        // Use the original key if name is null, otherwise use the name
-        const fieldName = formData[key].name === null ? key : formData[key].name
-        
-        // Handle JSON type values - parse them from string to object
-        let value = formData[key].value
-        if (formData[key].type === 'Json' && typeof value === 'string') {
-          try {
-            value = JSON.parse(value)
-          } catch (e) {
-            console.error(`Failed to parse JSON for ${fieldName}:`, value, e)
-            // Keep original value if parsing fails
-          }
-        }
-        
-        convertedFormData[fieldName] = value
-      })
-      
-      return convertedFormData
-    },
     showDiagram: function () {
       this.$refs.process.show()
       //TODO: Review b-modal static
